refactor: add resolveCommand helper and use common functions in install scripts
Introduces resolveCommand to handle commands installed to non-standard paths
(e.g. ~/.atuin/bin/atuin, ~/.local/bin, ~/.cargo/bin). All install scripts
now import commandExists/resolveCommand from common/functions.mjs instead of
duplicating inline `which` checks. Also sources atuin env in .zshrc and fixes
a dpkg-query template literal.

+++ linux/dotfiles/git/install.mjs
@@ -1,8 +1,8 @@
 #!/usr/bin/env zx
 
-try {
-  await $`which rebase-editor`;
-} catch {
+import { commandExists } from '../../../common/functions.mjs';
+
+if (!(await commandExists('rebase-editor'))) {
   console.log('  Installing rebase-editor');
   await $`npm install -g rebase-editor`;
   await $`git config --global sequence.editor rebase-editor`;
