+++ linux/dotfiles/vim/install.mjs
@@ -1,8 +1,8 @@
 #!/usr/bin/env zx
 
-try {
-  await $`which vim`;
-} catch {
+import { commandExists } from '../../../common/functions.mjs';
+
+if (!(await commandExists('vim'))) {
   console.log('  Installing vim for you.');
   await $`sudo apt-get install -y vim`;
 }
