+++ linux/dotfiles/vscode/install.mjs
@@ -1,8 +1,8 @@
 #!/usr/bin/env zx
 
-try {
-  await $`which code`;
-} catch {
+import { commandExists } from '../../../common/functions.mjs';
+
+if (!(await commandExists('code'))) {
   console.log('  Installing Visual Studio Code');
   await $`sudo snap install code --classic`;
 }
